requirements.txt
insightface
onnxruntime
opencv-python-headless
qdrant-client[grpc]
streamlit
pillow
numpy

